+++ src/datakit/__init__.py
@@ -10,5 +10,5 @@ from datakit.core.datakit import DataKit
 from datakit.config import config
 import datakit.safety.safe_ops as safe
 
-__version__ = "0.2.2"
+__version__ = "0.3.0"
 __all__ = ["DataKit", "config", "safe"]